MCTF - Writeup/Cute n Wonder/src/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.10
